Split the sync section into a directory with per-backend forms
