+++ src/scripts/options/sync-section.ts
@@ -1,0 +1,1 @@
+export * from "./sync-section/index.tsx";
